elf: provide method to access known symbols in an object
This function does not guarantee to return all the symbols in an object, as
that may be hard or impossible to compute. Instead it only promises to return
a subset of all symbols.

For objects which have section headers (such as non-stripped binaries), this set
should be complete though, because we register all symbols from the section headers
in the ELF class `__init__`.